Rename NavigationDrawer to TermListNavigationDrawer
